requirements.txt
Django==3.0.8
django-cleanup==5.0.0
djangorestframework==3.11.0
openpyxl==3.0.4
Pillow==7.2.0
psycopg2==2.8.5